libportable have getTimef since r1530 which gives a float instead of double, using it instead of mixing floats and doubles all over the place. It's not like double precision is needed anyways
svn path=/trunk/; revision=929

+++ src/state/TransitionState.cpp
@@ -23,10 +23,10 @@
 #include "TransitionState.h"
 #include "ViewState.h"
 
-double TransitionState::transition_time = 1.0;
+float TransitionState::transition_time = 1.0f;
 
 State* TransitionState::action(bool &flip){
-	double s = age() / transition_time;
+	float s = age() / transition_time;
 
 	gfx()->render( s );
 	flip = true;
